test: combine related test suites
